+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add `extractText` function to the Eyes class
 
 
 ## 12.7.0 - 2020/12/1